+++ nutritionlib/src/main/java/com/graydyn/nutritionlib/model/OcrPassData.kt
@@ -4,7 +4,7 @@ import android.graphics.Rect
 
 data class RawOcrLine(val text: String, val boundingBox: Rect?)
 
-data class MacroDetection(val macro: String, val value: Int, val fromLine: String)
+data class MacroDetection(val macro: String, val value: Float, val fromLine: String)
 
 data class OcrPassData(
     val rawLines: List<RawOcrLine>,       // all lines sorted top-to-bottom before row grouping
